Major code refactoring: Add NodeBetweennessInfo.
Added a data structure called NodeBetweennessInfo which stores
many of the variables needed for the calculation of
betweenness centrality on a per-node basis. Now the
computeAll() method of the GraphAnalyzers returns a hash map
mapping each node to its corresponding NodeBetweennessInfo.
This will facilitate the recovery of various information
(betweenness, closeness, etc.) for each node, which will
be especially usefull in GDMS-Topology.